// src/retailers/woolworths/dtos/woolworths-product.dto.ts
export class WoolworthsProductDto {
  AgeRestricted: boolean;
  Barcode: string;
  Brand: string;
  CupMeasure: string;
  CupPrice: number;
  Cupstring: string;
  Description: string;
  DetailsImagePaths: string[];
  Diagnostics: string;
  DisplayQuantity: number;
  FullDescription: string;
  GtinFormat: number;
  HasBeenBoughtBefore: boolean;
  HasCupPrice: boolean;
  HasHeaderTag: boolean;
  HasProductSubs: boolean;
  HideWasSavedPrice: boolean;
  InstoreCupPrice: number;
  InstoreCupstring: string;
  InstoreHasCupPrice: boolean;
  InstoreIsAvailable: boolean;
  InstoreIsOnSpecial: boolean;
  InstoreIsPurchasable: boolean;
  InstorePrice: number;
  InstoreSavingsAmount: number;
  InstoreWasPrice: number;
  IsAvailable: boolean;
  IsBundle: boolean;
  IsCentreTag: boolean;
  IsDeliveryPass: boolean;
  IsEdrSpecial: boolean;
  IsFooterEnabled: boolean;
  IsForCollection: boolean;
  IsForDelivery: boolean;
  IsForExpress: boolean;
  IsInFamily: boolean;
  IsInStock: boolean;
  IsInTrolley: boolean;
  IsMarketProduct: boolean;
  IsNew: boolean;
  IsOnSpecial: boolean;
  IsPmDelivery: boolean;
  IsPurchasable: boolean;
  IsRanged: boolean;
  IsRestrictedByDeliveryMethod: boolean;
  IsSponsoredAd: boolean;
  LargeImageFile: string;
  MediumImageFile: string;
  MinimumQuantity: number;
  Name: string;
  PackageSize: string;
  Price: number;
  QuantityInTrolley: number;
  RichDescription: string;
  SavingsAmount: number;
  SmallFormatDescription: string;
  SmallImageFile: string;
  Source: string;
  Stockcode: number;
  SupplyLimit: number;
  SupplyLimitMessage: string;
  Unit: string;
  UnitWeightInGrams: number;
  UrlFriendlyName: string;
  Variety: string;
  WasPrice: number;
}